Feature/or 3.0 (#201)
* Create Structured Data Object Actions

* OR Migration feature from XMLtoYAML

* updated sharedorlocation

* fixed shortcut keys for ccp

* added highlight for pasted object/page

* fixed move to shared option

* fixed removeunusedobjects

* Fix OR changes (#167)

* Fix empty pages and delete object not reflecting in YAML

* Fix and enhanced get impacted test cases

* Fix Delete option both proj and shared not reflecting in backend

* Fix reference scope of object in test cases when move to shared

* Fix [Porject] reference is not added into the test case when Scenario name has leading Spaces

* Fix issue in Reference column for object Project/Scope is gone when clicking in the Test case UI

* Enhance move functionality for WebOR and MobileOR pages to shared repository

* Add page name extraction for scoped references in navigateToObject method

* Fix go to reusables obj test data (#171)

* Enhance scenario retrieval logic to prioritize reusable scenarios before falling back to regular scenarios

* Add page name extraction for scoped references in navigateToObject method

* Refactor navigateToObject method to support scoped navigation for shared and project trees

* fix after merge plugins

* StructuredData OR Updates

* addressed comments

* added space as recommended best practice by git

* Highlight OR icons when selected

* Add new SAP Testing Feature (#173)

* OR Migration feature from XMLtoYAML

* Add new SAP Feature in 3.0

* address comments

* revert deleted code in PlaywrigthRecorderParser

* added missing scope in yaml format

---------


---------

+++ Datalib/src/main/java/com/ing/datalib/or/structureddata/StructuredDataORObject.java
@@ -1,7 +1,6 @@
-package com.ing.datalib.or.api;
+package com.ing.datalib.or.structureddata;
 
 import com.ing.datalib.component.utils.FileUtils;
-import com.ing.datalib.or.common.ORAttribute;
 import com.ing.datalib.or.common.ORObjectInf;
 import com.ing.datalib.or.common.ORUtils;
 import com.ing.datalib.or.common.ObjectGroup;
@@ -28,36 +27,35 @@ import javax.swing.tree.TreePath;
  * Supports only JsonPath and Xpath as locator attributes.
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class APIORObject extends UndoRedoModel implements ORObjectInf {
+public class StructuredDataORObject extends UndoRedoModel implements ORObjectInf {
 
     @JacksonXmlProperty(isAttribute = true, localName = "ref")
     private String name;
 
     @JacksonXmlProperty(localName = "Property")
     @JacksonXmlElementWrapper(useWrapping = false, localName = "Property")
-    private List<ORAttribute> attributes;
+    private List<StructuredDataAttribute> attributes;
 
     @JsonIgnore
-    private ObjectGroup<APIORObject> group;
+    private ObjectGroup<StructuredDataORObject> group;
 
-    public APIORObject() {
-        setDefaultORAttributes();
+    public StructuredDataORObject() {
+        setDefaultStructuredDataAttributes();
     }
 
-    public APIORObject(String name, ObjectGroup group) {
+    public StructuredDataORObject(String name, ObjectGroup group) {
         this.name = name;
         this.group = group;
-        setDefaultORAttributes();
+        setDefaultStructuredDataAttributes();
     }
 
     @JsonIgnore
-    public final void setDefaultORAttributes() {
+    public final void setDefaultStructuredDataAttributes() {
         attributes = new ArrayList<>();
-        for (int i = 0; i < APIOR.OBJECT_PROPS.size(); i++) {
-            ORAttribute attr = new ORAttribute();
-            attr.setName(APIOR.OBJECT_PROPS.get(i));
+        for (int i = 0; i < StructuredDataOR.OBJECT_PROPS.size(); i++) {
+            StructuredDataAttribute attr = new StructuredDataAttribute();
+            attr.setName(StructuredDataOR.OBJECT_PROPS.get(i));
             attr.setValue("");
-            attr.setPreference("" + (i + 1));
             attributes.add(attr);
         }
     }
@@ -72,11 +70,11 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
         this.name = name;
     }
 
-    public List<ORAttribute> getAttributes() {
+    public List<StructuredDataAttribute> getAttributes() {
         return attributes;
     }
 
-    public void setAttributes(List<ORAttribute> attributes) {
+    public void setAttributes(List<StructuredDataAttribute> attributes) {
         this.attributes = attributes;
     }
 
@@ -88,7 +86,9 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
             group.removeFromParent();
         }
         group.getObjects().remove(this);
-        FileUtils.deleteFile(getRepLocation());
+        if (!group.getParent().getRoot().getObjectRepository().isUsingYamlFormat()) {
+            FileUtils.deleteFile(getRepLocation());
+        } 
     }
 
     @JsonIgnore
@@ -105,7 +105,7 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
 
     @JsonIgnore
     @Override
-    public ObjectGroup<APIORObject> getParent() {
+    public ObjectGroup<StructuredDataORObject> getParent() {
         return group;
     }
 
@@ -170,7 +170,7 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
     @JsonIgnore
     @Override
     public void setValueAt(Object value, int rowIndex, int columnIndex) {
-        ORAttribute attr = attributes.get(rowIndex);
+        StructuredDataAttribute attr = attributes.get(rowIndex);
         if (columnIndex == 0) {
             if (isNotDefault(rowIndex) && getAttribute(value.toString()) == null) {
                 super.setValueAt(value, rowIndex, columnIndex);
@@ -189,7 +189,7 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
     @JsonIgnore
     private Boolean isNotDefault(int rowIndex) {
         String value = getValueAt(rowIndex, 0).toString();
-        return APIOR.OBJECT_PROPS.indexOf(value) == -1;
+        return StructuredDataOR.OBJECT_PROPS.indexOf(value) == -1;
     }
 
     @JsonIgnore
@@ -212,13 +212,13 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
     @JsonIgnore
     private void changeSave() {
         if (group != null) {
-            APIORPage page = (APIORPage) group.getParent();
+            StructuredDataORPage page = (StructuredDataORPage) group.getParent();
             page.getRoot().setSaved(false);
             
             // Auto-save for YAML format
             if (page.getRoot().getObjectRepository() != null 
                 && page.getRoot().getObjectRepository().isUsingYamlFormat()) {
-                page.getRoot().getObjectRepository().saveAPIPageNow(page);
+                page.getRoot().getObjectRepository().saveStructuredDataPageNow(page);
             }
         }
     }
@@ -279,8 +279,8 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
 
     @JsonIgnore
     @Override
-    public APIORPage getPage() {
-        return (APIORPage) group.getParent();
+    public StructuredDataORPage getPage() {
+        return (StructuredDataORPage) group.getParent();
     }
 
     @JsonIgnore
@@ -310,28 +310,18 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
     @JsonIgnore
     @Override
     public Boolean rename(String newName) {
-        Boolean flag = true;
         if (getParent().getChildCount() == 1) {
-            flag = getParent().rename(newName);
+            getParent().rename(newName);
         }
-        if (flag && getParent().getObjectByName(newName) == null) {
-            // Check if using YAML format
-            if (getParent().getParent().getRoot().getObjectRepository().isUsingYamlFormat()) {
-                // For YAML format, objects are stored within the page YAML file
-                // Just update the name and mark as needing save
-                setName(newName);
-                changeSave();
-                return true;
-            } else {
-                // Use original XML folder-based rename
-                if (FileUtils.renameFile(getRepLocation(), newName)) {
-                    setName(newName);
-                    changeSave();
-                    return true;
-                }
-            }
+        if (newName == null || newName.isBlank()) {
+            return false;
         }
-        return false;
+        if (getParent().getObjectByName(newName) != null) {
+            return false;
+        }
+        setName(newName);
+        changeSave();
+        return true;
     }
 
     @JsonIgnore
@@ -342,11 +332,11 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
 
     @JsonIgnore
     @Override
-    public APIORObject clone(ORObjectInf obj) {
-        if (obj instanceof APIORObject) {
-            APIORObject apiObj = (APIORObject) obj;
+    public StructuredDataORObject clone(ORObjectInf obj) {
+        if (obj instanceof StructuredDataORObject) {
+            StructuredDataORObject apiObj = (StructuredDataORObject) obj;
             apiObj.getAttributes().clear();
-            for (ORAttribute attribute : attributes) {
+            for (StructuredDataAttribute attribute : attributes) {
                 apiObj.getAttributes().add(attribute.cloneAs());
             }
             apiObj.changeSave();
@@ -356,8 +346,8 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
     }
 
     @JsonIgnore
-    public ORAttribute getAttribute(String name) {
-        for (ORAttribute attribute : attributes) {
+    public StructuredDataAttribute getAttribute(String name) {
+        for (StructuredDataAttribute attribute : attributes) {
             if (attribute.getName().equalsIgnoreCase(name)) {
                 return attribute;
             }
@@ -367,7 +357,7 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
 
     @JsonIgnore
     public String getAttributeByName(String name) {
-        for (ORAttribute attribute : attributes) {
+        for (StructuredDataAttribute attribute : attributes) {
             if (attribute.getName().equalsIgnoreCase(name)) {
                 return attribute.getValue();
             }
@@ -377,7 +367,7 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
 
     @JsonIgnore
     public void setAttributeByName(String name, String value) {
-        for (ORAttribute attribute : attributes) {
+        for (StructuredDataAttribute attribute : attributes) {
             if (attribute.getName().equalsIgnoreCase(name)) {
                 attribute.setValue(value);
             }
@@ -386,16 +376,15 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
 
     @JsonIgnore
     public void addOrUpdateAttribute(String name, String value) {
-        for (ORAttribute attribute : attributes) {
+        for (StructuredDataAttribute attribute : attributes) {
             if (attribute.getName().equalsIgnoreCase(name)) {
                 attribute.setValue(value);
                 return;
             }
         }
-        ORAttribute attr = new ORAttribute();
+        StructuredDataAttribute attr = new StructuredDataAttribute();
         attr.setName(name);
         attr.setValue(value);
-        attr.setPreference(String.valueOf(attributes.size() + 1));
         attributes.add(attr);
         changeSave();
     }
@@ -409,21 +398,32 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
             name = attrName + i++;
         } while (getAttribute(name) != null);
         
-        ORAttribute attr = new ORAttribute();
+        StructuredDataAttribute attr = new StructuredDataAttribute();
         attr.setName(name);
         attr.setValue("");
-        attr.setPreference(String.valueOf(attributes.size() + 1));
         attributes.add(attr);
         changeSave();
         fireTableRowsInserted(attributes.size() - 1, attributes.size() - 1);
     }
 
     @JsonIgnore
-    public void removeAttribute(int attrIndex) {
-        if (attrIndex >= 0 && attrIndex < attributes.size()) {
-            attributes.remove(attrIndex);
-            changeSave();
-            fireTableRowsDeleted(attrIndex, attrIndex);
+    public void addNewAttribute(String attrName) {
+        if (getAttribute(attrName) == null) {
+            attributes.add(new StructuredDataAttribute(attrName, attributes.size()));
+            super.rowAdded(attributes.size() - 1);
+            fireTableRowsInserted(attributes.size() - 1, attributes.size() - 1);
+        }
+    }
+
+    @JsonIgnore
+    public void removeAttribute(String attrName) {
+        if (StructuredDataOR.OBJECT_PROPS.indexOf(attrName) == -1) {
+            if (getAttribute(attrName) != null) {
+                int index = attributes.indexOf(getAttribute(attrName));
+                super.rowDeleted(index);
+                attributes.remove(index);
+                fireTableRowsDeleted(index, index);
+            }
         }
     }
 
@@ -442,14 +442,13 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
             return;
         }
         
-        ORAttribute attr = new ORAttribute();
+        StructuredDataAttribute attr = new StructuredDataAttribute();
         if (values != null && values.length > 0) {
             attr.setName(values[0] != null ? values[0].toString() : "");
             if (values.length > 1) {
                 attr.setValue(values[1] != null ? values[1].toString() : "");
             }
         }
-        attr.setPreference(String.valueOf(row + 1));
         
         attributes.add(row, attr);
         rowAdded(row);
@@ -459,10 +458,10 @@ public class APIORObject extends UndoRedoModel implements ORObjectInf {
     @JsonIgnore
     @Override
     public Boolean isEqualOf(ORObjectInf object) {
-        if (object == null || !(object instanceof APIORObject)) {
+        if (object == null || !(object instanceof StructuredDataORObject)) {
             return false;
         }
-        APIORObject other = (APIORObject) object;
+        StructuredDataORObject other = (StructuredDataORObject) object;
         return Objects.equals(this.name, other.name);
     }
